requirements.txt
Flask
python-dotenv
redis
flask_sqlalchemy